+++ examples/with-markprompt-web/src/main.ts
@@ -8,5 +8,6 @@ const el = document.querySelector('#markprompt');
 if (el && el instanceof HTMLElement) {
   markprompt(import.meta.env.VITE_PROJECT_API_KEY, el, {
     trigger: { floating: true },
+    search: { enable: true },
   });
 }
