+++ packages/react/src/primitives/footer.test.tsx
