+++ packages/react/src/primitives/headless.tsx
@@ -1,4 +1,3 @@
-import type { SubmitPromptOptions } from '@markprompt/core';
 import * as Dialog from '@radix-ui/react-dialog';
 import debounce from 'p-debounce';
 import React, {
@@ -7,10 +6,9 @@ import React, {
   useEffect,
   useMemo,
   useRef,
-  useState,
   type ChangeEventHandler,
-  type ComponentPropsWithoutRef,
   type ComponentPropsWithRef,
+  type ComponentPropsWithoutRef,
   type ElementType,
   type FormEventHandler,
   type KeyboardEventHandler,
@@ -22,16 +20,15 @@ import Markdown from 'react-markdown';
 import { mergeRefs } from 'react-merge-refs';
 import remarkGfm from 'remark-gfm';
 
-import { ConditionalVisuallyHidden } from './ConditionalWrap.js';
-import { MarkpromptContext, useMarkpromptContext } from './context.js';
 import { Footer } from './footer.js';
+import { ConditionalVisuallyHidden } from '../ConditionalWrap.js';
+import { MarkpromptContext, useMarkpromptContext } from '../context.js';
 import type {
   FlattenedSearchResult,
   PolymorphicComponentPropWithRef,
   PolymorphicRef,
-} from './types.js';
-import { useMarkprompt } from './useMarkprompt.js';
-import { type UseMarkpromptOptions } from './useMarkprompt.js';
+} from '../types.js';
+import { useMarkprompt, type UseMarkpromptOptions } from '../useMarkprompt.js';
 
 type RootProps = Dialog.DialogProps & UseMarkpromptOptions;
