+++ packages/react/src/primitives/index.tsx
@@ -1,10 +1,4 @@
 export {
-  useMarkprompt,
-  type UseMarkpromptOptions,
-  type UseMarkpromptResult,
-} from './useMarkprompt.js';
-
-export {
   Answer,
   AutoScroller,
   Close,
@@ -35,10 +29,6 @@ export {
   type SearchResultProps,
   type SearchResultsProps,
   type TitleProps,
-} from './primitives/headless.js';
+} from './headless.js';
 
-export { Markprompt } from './Markprompt.js';
-
-export { useMarkpromptContext } from './context.js';
-
-export * from './types.js';
+export { Footer } from './footer.js';
